+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Release History
 
+## 1.1.0 (...)
+- major refactoring of the code
+- API change for authentication plugins, not for consumers of this API
+- no longer support multiple authentication methods on one endpoint, it
+  is not used anywhere
+
 ## 1.0.2 (2015-10-13)
 - fix small bug when mentioned friendly name of authentication plugin 
   does not exist